ai_be/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-community>=0.0.10
transformers>=4.36.0
huggingface-hub>=0.20.1
sentence-transformers>=2.2.2